move epic styles to a wrapper component
